Finished moving webview logic to separate component
